requirements.txt
unstructured[pdf]
pillow
lxml
chromadb
tiktoken
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
pymupdf
pydantic==2.10.5
bitsandbytes
pymongo
ragas
langfuse
rank_bm25